found 7 days forecast call. Functionality, except of geoposition is done

+++ app/src/main/java/com/nieduard/weather_avito/model/Lst.kt
@@ -11,7 +11,7 @@ data class Lst(
     val feels_like: FeelsLike,
     val gust: Double,
     val humidity: Int,
-    val pop: Int,
+    val pop: Double,
     val pressure: Int,
     val speed: Double,
     val sunrise: Int,
